#713 get rid of Utils class, integrate URL building logic in LinkHandler class
