+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/internal/Heading.java
@@ -1,0 +1,48 @@
+/*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+ ~ Copyright 2017 Adobe
+ ~
+ ~ Licensed under the Apache License, Version 2.0 (the "License");
+ ~ you may not use this file except in compliance with the License.
+ ~ You may obtain a copy of the License at
+ ~
+ ~     http://www.apache.org/licenses/LICENSE-2.0
+ ~
+ ~ Unless required by applicable law or agreed to in writing, software
+ ~ distributed under the License is distributed on an "AS IS" BASIS,
+ ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ ~ See the License for the specific language governing permissions and
+ ~ limitations under the License.
+ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
+package com.adobe.cq.wcm.core.components.internal;
+
+import org.apache.commons.lang3.StringUtils;
+
+public enum Heading {
+
+    H1("h1"),
+    H2("h2"),
+    H3("h3"),
+    H4("h4"),
+    H5("h5"),
+    H6("h6");
+
+    private String element;
+
+    Heading(String element) {
+        this.element = element;
+    }
+
+    public static Heading getHeading(String value) {
+        for (Heading heading : values()) {
+            if (StringUtils.equalsIgnoreCase(heading.element, value)) {
+                return heading;
+            }
+        }
+        return null;
+    }
+
+    public String getElement() {
+        return element;
+    }
+
+}
